# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "REST API that allows clients to communicate with drones (i.e. **dispatch controller**)",
        "title": "github template restapi",
        "contact": {
            "name": "Daniel Mena and Kmilo Denis Glez",
            "url": "https://github.com/dani-fmena and https://github.com/kmilodenisglez",
            "email": "kmilo.denis.glez@gmail.com"
        },
        "version": "0.1"
    },
    "basePath": "/",
    "paths": {
        "/auth": {
            "post": {
                "description": "Intent to grant authentication using the provider user's credentials and the specified  auth provider\r\n\r\nUser Credentials:\r\n\r\n|  Username   | Password    | \r\n| ----------- | -----------|\r\n| richard.sargon@meinermail.com | password1 |\r\n| tom.carter@meinermail.com | password2 |\r\n",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Auth"
                ],
                "summary": "User authentication",
                "parameters": [
                    {
                        "description": "User Login Credential",
                        "name": "credential",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.UserCredIn"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "err.wrong_auth_provider",
                        "schema": {
                            "$ref": "#/definitions/dto.Problem"
                        }
                    },
                    "401": {
                        "description": "err.unauthorized",
                        "schema": {
                            "$ref": "#/definitions/dto.Problem"
                        }
                    },
                    "500": {
                        "description": "err.json_parse",
                        "schema": {
                            "$ref": "#/definitions/dto.Problem"
                        }
                    },
                    "504": {
                        "description": "err.network",
                        "schema": {
                            "$ref": "#/definitions/dto.Problem"
                        }
                    }
                }
            }
        },
        "/auth/logout": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "This endpoint invalidated a previously granted access token",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Auth"
                ],
                "summary": "User logout",
                "parameters": [
                    {
                        "type": "string",
                        "default": "Bearer \u003cAdd access token here\u003e",
                        "description": "Insert access token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "OK"
                    },
                    "401": {
                        "description": "err.unauthorized",
                        "schema": {
                            "$ref": "#/definitions/dto.Problem"
                        }
                    },
                    "500": {
                        "description": "err.generic",
                        "schema": {
                            "$ref": "#/definitions/dto.Problem"
                        }
                    }
                }
            }
        },
        "/auth/user": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Auth"
                ],
                "parameters": [
                    {
                        "type": "string",
                        "default": "Bearer \u003cAdd access token here\u003e",
                        "description": "Insert access token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/dto.User"
                            }
                        }
                    },
                    "401": {
                        "description": "err.unauthorized",
                        "schema": {
                            "$ref": "#/definitions/dto.Problem"
                        }
                    },
                    "500": {
                        "description": "err.generic",
                        "schema": {
                            "$ref": "#/definitions/dto.Problem"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "dto.Problem": {
            "type": "object",
            "properties": {
                "detail": {
                    "type": "string",
                    "example": "Some error details"
                },
                "status": {
                    "type": "integer",
                    "example": 503
                },
                "title": {
                    "type": "string",
                    "example": "err_code"
                }
            }
        },
...
        "dto.User": {
            "type": "object",
            "properties": {
                "name": {
                    "type": "string"
                },
                "passphrase": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        },
        "dto.UserCredIn": {
            "type": "object",
            "required": [
                "password",
                "username"
            ],
            "properties": {
                "password": {
                    "type": "string",
                    "example": "password1"
                },
                "username": {
                    "type": "string",
                    "example": "richard.sargon@meinermail.com"
                }
            }
        }
    }
}